Character movement scales wit dilated time.
