+++ wecs/mechanics/__init__.py
@@ -1,5 +1,3 @@
 from .clock import SettableClock
 from .clock import Clock
 from .clock import DetermineTimestep
-from .stamina import Stamina
-from .stamina import UpdateStamina
